Move all service related declarations under monaco.services
